rag/requirements.txt
streamlit
pdfplumber
python-docx
google-generativeai
sentence-transformers
chromadb
